moved driver.java to user registration package
