+++ AuPairPlacementSystem/src/main/java/edu/srh/aupair/utilities/utilities.java
@@ -16,7 +16,8 @@ public class utilities {
 
 	}
 
-	public static Connection getConnectionString() throws SQLException {
+	public static Connection getConnectionString() throws SQLException 
+	{
 		Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/AU_PAIR_MANAGEMENT?useSSL=false", "root",
 				"Pass123$$");
 		return conn;
